enhance: [satellite] provide the way to specify subscription by its contracts
ADD_SUBSCRIPTION_TO_ACTIVATION_KEYS in config.sh uses a subscription
name as a pattern for extracting a subscription id from the output of
hammer command. The pattern is embedded into an argument for sed command.

However, writing a subscription name as a effective pattern for extracting
in yml is not easy because:

* a subscription name can include comma characters and parenthesis characters like:

  Red Hat Enterprise Linux Server, Standard (Physical or Virtual Nodes)

* csv output of hammer users double quotes for warpping a subscription name like:

    # hammer --csv subscription list
    ID,UUID,Name,Type,Contract,Account,Support,End Date,Quantity,Consumed
    1,xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx,"Red Hat Enterprise Linux Server, Premium (...)",...

This change provides the way to specify MULTIPLE contracts for
extracting a subscription ids instead of names.

"rct cat-manifest" command helps to know the available contracts from a manifest
zip file.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>