enhance: use python's print as a function in vmbuild
In vmbuild.sh script, python command is used for generating a
mac address. However, the script assumes implicitly that the python
command is python2; print is not a function. As the result, the
script doesn't work well on Fedora 31 where python command is
python3.

In python3, print is a function. It takes arguments surrounded by
`(' and `)'.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>