enhance: [rhel7-min] use "inst." prefix even when a rhel7 variant is specified
On Fedora34 as an build host, it seems that virtinst.os_variant,
which is passed to virt-isntall ... --os-variant=, must be listed
in $(osinfo-query os).

About rhel7, $(osinfo-query os) lists only "rhel7-unknown", "rhel7.0",
"rhel7.1",..., "rhel7.9; no "rhel7" is in the list anymore.

The original code expected only "rhel7" when expanding
extra_args in vmbuild.sh.

With this change, "rhel7*" variants are acceptable.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>